+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Ark-Lock
 
+## 0.3
+
+Multi-Mutex-Lock Support
+
+## 0.2
+
+Put under Packagist for Composer
+
 ## 0.1
 
 Framework and Basic Support for FLOCK, MYSQL, and a DUMMY.